+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Değişiklik Günlüğü
 
+## [Yayımlanmadı]
+
+### Eklendi
+- Frekans analizi aracı
+- Sezar şifresi kırma aracı
+
+### Değiştirildi
+- Güncel geliştirici bilgisi yalnızca Mehmet Zahid KAYA olarak düzenlendi.
+
 ## [1.1.0] - 2026-08-05
 
 ### Eklendi